requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
neo4j==5.17.0
redis==5.0.1

# AI/ML
google-generativeai==0.4.0
tiktoken==0.6.0
numpy==1.26.4
dspy>=2.5.0

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.3
feedparser==6.0.11

# PDF Processing
pymupdf>=1.24.0

# Data Processing
pydantic==2.6.1
pydantic-settings==2.1.0
python-dateutil==2.8.2
beautifulsoup4==4.12.3

# Object Storage
minio==7.2.3

# Background Tasks
celery==5.3.6
apscheduler==3.10.4

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
click==8.1.7

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
faker==22.6.0

# Development
black==24.1.1
ruff==0.2.1
mypy==1.8.0

# MedGemma (Google HAI-DEF) - Required for MedGemma Impact Challenge
# Models: google/medgemma-1.5-4b-it, google/medgemma-4b-it, google/medgemma-27b-it
transformers>=4.47.0  # HuggingFace transformers for MedGemma
accelerate>=1.0.0  # Model loading and device management
bitsandbytes>=0.44.0  # 4-bit quantization for lower memory
huggingface_hub>=0.27.0  # Model download
torch>=2.0.0  # PyTorch backend

# Medical Literature (MedLit Agent)
chromadb>=0.5.0  # Local vector store (uses newer hnswlib)
biopython>=1.83  # PubMed API (Entrez)
sentence-transformers>=2.0.0  # PubMedBERT embeddings

# Full Stack Healthcare Platform (Patient + Admin layers)
twilio>=8.0.0  # SMS appointment reminders & Voice AI
python-jose[cryptography]>=3.3.0  # JWT authentication
passlib[bcrypt]>=1.7.4  # Password hashing
bcrypt>=4.0.0  # Password hashing backend
